pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-my-env-v4"
version = "1.0.0"
description = "OpenEnv content moderation benchmark with typed OpenEnv client/server contracts."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Krishna Ramasamy" }
]
dependencies = [
  "openenv-core>=0.1.1",
  "openai>=1.40.0,<2.0.0",
  "requests>=2.31.0,<3.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0"
]

[tool.setuptools]
py-modules = ["models", "client", "inference"]

[tool.setuptools.packages.find]
include = ["server"]